Изменён тип переменной

+++ steam_web/API/Models/IPlayerService/PlayerOwnedGames.cs
@@ -3,6 +3,6 @@
 namespace SteamWeb.API.Models.IPlayerService;
 public class PlayerOwnedGames
 {
-    [JsonPropertyName("game_count")] public uint GameCount { get; init; }
+    [JsonPropertyName("game_count")] public ushort GameCount { get; init; }
     [JsonPropertyName("games")] public OwnedGame[] Games { get; init; } = Array.Empty<OwnedGame>();
 }
